refactor: address PR #2553 review feedback (AN-BC)
Two pieces of review feedback:

1. **Move `POSTGRESQL_PROCEDURE_PARAM_LIMIT` to the PG provider package**.
   Was in `@memberjunction/generic-database-provider` because the rules
   module lived there. Reviewer correctly noted it's a PG-specific
   constant and belongs in `@memberjunction/postgresql-dataprovider`
   alongside the `ProcedureParamLimit` getter that exposes it.

   The rules module (`useJsonArgShape`, `projectedParamCount`) takes the
   limit as a `paramLimit` argument — it never imported the constant
   directly — so the move is callsite-only:

   - Removed: `crudSprocFieldRules.ts` definition + index re-export
   - Added: `PostgreSQLDataProvider.ts` definition + index export
   - Updated import in `PostgreSQLCodeGenProvider.ts` to pull from
     `@memberjunction/postgresql-dataprovider`

2. **Collapse the hardcoded `if (platform === 'postgresql')` blocks in
   `manage-metadata.ts` and `sql.ts` to a unified
   `ClassFactory.CreateInstance` call**. Reviewer's exact wording: "Why
   fall back to SQL and why hardcode check for PG provider and then use
   class factory for it. seems odd."

   Both `SQLServerCodeGenProvider` and `PostgreSQLCodeGenProvider` are
   decorated with `@RegisterClass(CodeGenDatabaseProvider, '<platform>')`,
   so a single platform-keyed lookup resolves the right provider for
   either platform. Replaced the if/else with:

       const provider = MJGlobal.Instance.ClassFactory
           .CreateInstance(CodeGenDatabaseProvider, platform);
       if (!provider || provider.constructor === CodeGenDatabaseProvider) {
           throw new Error(`CodeGen provider for dbPlatform='${platform}' not found...`);
       }

   The `constructor === CodeGenDatabaseProvider` guard catches the
   silent-fallback-to-abstract-base path that motivated the original
   provider-lookup-key fix in the parent commit.

   `import { SQLServerCodeGenProvider }` replaced with side-effect-only
   `import './providers/sqlserver/SQLServerCodeGenProvider'` — the import
   is still needed to trigger the `@RegisterClass` decorator side effect
   that registers the class with `ClassFactory`. Without the import,
   `ClassFactory.CreateInstance(..., 'sqlserver')` would return null.

## Verification

- Build: clean across `CodeGenLib`, `PostgreSQLDataProvider`,
  `GenericDatabaseProvider`.
- Tests: 447 / 120 / 503 passing (1,070 total).
- Fresh DB: 128 migrations apply cleanly; Bug 5 sproc returns 4,455
  rows; JSON-arg `spCreateAIAgent` works end-to-end with default
  application (`Status='Pending'`).
- Real `mj codegen` run against the verify DB succeeds (the unified
  `ClassFactory.CreateInstance` lookup resolves correctly for the
  configured `'postgresql'` platform).
