+++ packages/PostgreSQLDataProvider/src/index.ts
@@ -1,4 +1,4 @@
-export { PostgreSQLDataProvider } from './PostgreSQLDataProvider.js';
+export { PostgreSQLDataProvider, POSTGRESQL_PROCEDURE_PARAM_LIMIT } from './PostgreSQLDataProvider.js';
 export { PGConnectionManager, PGConnectionConfig } from './pgConnectionManager.js';
 export { PGQueryParameterProcessor } from './queryParameterProcessor.js';
 export { PostgreSQLProviderConfigData, PostgreSQLProviderConfigOptions } from './types.js';
